+++ src/MobileChat.MAUI/Pages/SettingsPage.razor
@@ -1,7 +1,7 @@
 ﻿@page "/settings"
 
 @using MobileChat.MAUI.Interfaces
-@using MobileChat.MAUI.Models
+@using MobileChat.MAUI.Services
 @using MobileChat.MAUI.Shared.Components
 @using MobileChat.Shared.Interfaces
 @using MobileChat.Shared.Models
