+++ src/MobileChat.MAUI/Services/SessionStorage.cs
@@ -1,6 +1,6 @@
 ﻿using MobileChat.Shared.Models;
 
-namespace MobileChat.MAUI.Models
+namespace MobileChat.MAUI.Services
 {
     public class SessionStorage
     {
