+++ src/MobileChat.MAUI/Shared/Components/SignInView.razor
@@ -4,11 +4,11 @@
 @using System.Text.RegularExpressions
 @using MobileChat.MAUI.Helpers
 @using MobileChat.MAUI.Interfaces
-@using MobileChat.MAUI.Models
+@using MobileChat.MAUI.Services
 @using MobileChat.Shared.Interfaces
 @using MobileChat.Shared.Models
 @inject ISaveFile SaveFileService
-@inject IChatHub ChatService
+@inject IChatAuth ChatAuthService
 @inject SessionStorage SessionStorage
 @inject NavigationManager NavigationManager
 
@@ -77,7 +77,7 @@
             return;
         }
 
-        KeyValuePair<Guid, bool> values = await ChatService.SignIn(EmailOrUsername, Password);
+        KeyValuePair<Guid, bool> values = await ChatAuthService.SignIn(EmailOrUsername, Password);
 
         //check user
         if (values.Value)
