+++ src/MobileChat.MAUI/Shared/Components/SignUpView.razor
@@ -3,11 +3,12 @@
 @using FluentValidation
 @using System.Text.RegularExpressions
 @using MobileChat.MAUI.Interfaces;
-@using MobileChat.MAUI.Models
+@using MobileChat.MAUI.Services
 @using MobileChat.Shared.Interfaces
 @using MobileChat.Shared.Models
 @inject ISaveFile SaveFileService
-@inject IChatHub ChatService
+@inject IChatAuth ChatAuthService
+@inject IChatUser ChatUserService
 @inject SessionStorage SessionStorage
 @inject NavigationManager NavigationManager
 
@@ -74,7 +75,7 @@
 
     protected override void OnInitialized()
     {
-        UserValidator = new UserFluentValidator(ChatService);
+        UserValidator = new UserFluentValidator(ChatUserService);
 
         base.OnInitialized();
     }
@@ -92,7 +93,7 @@
             return;
         }
 
-        KeyValuePair<Guid, bool> values = await ChatService.SignUp(model.DisplayName, model.Username, model.Email, model.Password);
+        KeyValuePair<Guid, bool> values = await ChatAuthService.SignUp(model.DisplayName, model.Username, model.Email, model.Password);
 
         //create user
         if (values.Value)
@@ -142,10 +143,10 @@
 
     public class UserFluentValidator : AbstractValidator<User>
     {
-        private readonly IChatHub ChatService;
-        public UserFluentValidator(IChatHub ChatService)
+        private readonly IChatUser ChatUserService;
+        public UserFluentValidator(IChatUser ChatUserService)
         {
-            this.ChatService = ChatService;
+            this.ChatUserService = ChatUserService;
 
             RuleFor(x => x.DisplayName)
                 .NotEmpty()
