+++ src/MobileChat.MAUI/Shared/MainLayout.razor
@@ -1,7 +1,6 @@
 ﻿@using MobileChat.Client
-@using MobileChat.MAUI.Models
+@using MobileChat.MAUI.Services
 @inherits LayoutComponentBase
-
 @inject SessionStorage SessionStorage
 
 <MudThemeProvider @ref="@_mudThemeProvider" @bind-IsDarkMode="@_isDarkMode" />
